# api/docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Tomas API",
    "version": "2.0.0",
    "description": "\n        Personal and Home finance projection and expense tracking application API.\n        \n        ## Authentication\n        Most endpoints require JWT authentication. Include the token in the Authorization header:\n        ```\n        Authorization: Bearer <your-jwt-token>\n        ```\n        \n        ## Parameter Naming Convention\n        All API endpoints use `snake_case` for query parameters and request body fields.\n        \n        ## Breaking Changes (v2.0.0)\n        - Removed legacy `/transactions?book_id=...` endpoint - use `/books/{id}/transactions` instead\n        - All parameters now use `snake_case` (e.g., `account_id`, `start_date`, `end_date`, `up_to_date`)\n      ",
    "contact": {
      "name": "Tomas API Support",
      "url": "https://github.com/levhita/tomas"
    },
    "license": {
      "name": "GPL-3.0",
      "url": "https://www.gnu.org/licenses/gpl-3.0.html"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    },
    {
      "url": "https://api.tomas.example.com/api",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Account": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "book_id": {
            "type": "integer",
            "example": 123
          },
          "name": {
            "type": "string",
            "example": "Checking Account"
          },
          "note": {
            "type": "string",
            "nullable": true,
            "example": "Main checking account"
          },
          "type": {
            "type": "string",
            "enum": [
              "debit",
              "credit"
            ],
            "example": "debit"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:30:00Z"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "error"
          },
          "message": {
            "type": "string",
            "example": "Resource not found"
          },
          "code": {
            "type": "string",
            "example": "NOT_FOUND"
          }
        }
      }
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Authentication required",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "status": "error",
              "message": "Authentication required",
              "code": "UNAUTHORIZED"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Access forbidden",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "status": "error",
              "message": "Access forbidden",
              "code": "FORBIDDEN"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "status": "error",
              "message": "Resource not found",
              "code": "NOT_FOUND"
            }
          }
        }
      }
    }
  },
...
  "paths": {
    "/accounts/{id}": {
      "get": {
        "summary": "Get details for a single account",
        "description": "Retrieve details for a specific account by ID. Requires read access to the account's book.",
        "tags": [
          "Accounts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Account ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Account details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Account"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "description": "Failed to fetch account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/accounts/{id}/balance": {
      "get": {
        "summary": "Get current balance for an account, optionally up to a specific date",
        "description": "Calculate account balance with exercised and projected totals, optionally filtered by date.",
        "tags": [
          "Accounts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Account ID"
          },
          {
            "in": "query",
            "name": "up_to_date",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Optional date filter in YYYY-MM-DD format",
            "example": "2024-04-30"
          }
        ],
        "responses": {
          "200": {
            "description": "Account balance information",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "account_id": {
                      "type": "integer",
                      "example": 123
                    },
                    "exercised_balance": {
                      "type": "number",
                      "format": "float",
                      "example": 1500.5,
                      "description": "Balance from exercised transactions only"
                    },
                    "projected_balance": {
                      "type": "number",
                      "format": "float",
                      "example": 1200.75,
                      "description": "Balance including all transactions"
                    },
                    "up_to_date": {
                      "type": "string",
                      "format": "date",
                      "nullable": true,
                      "example": "2024-04-30"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "description": "Failed to calculate balance",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/books/{id}/accounts": {
      "get": {
        "summary": "Get all accounts for a specific book",
        "description": "Retrieve all accounts belonging to a specific book. Requires read access via team membership.",
        "tags": [
          "Books"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Book ID"
          }
        ],
        "responses": {
          "200": {
            "description": "List of accounts for the book",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Account"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "description": "Failed to fetch accounts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}